# Add support for infix wildcard